created url_parser and removed duplication
