+++ valkey/_parsers/__init__.py
@@ -4,7 +4,7 @@ from .encoders import Encoder
 from .hiredis import _AsyncHiredisParser, _HiredisParser
 from .resp2 import _AsyncRESP2Parser, _RESP2Parser
 from .resp3 import _AsyncRESP3Parser, _RESP3Parser
-
+from url_parser import parse_url
 __all__ = [
     "AsyncCommandsParser",
     "_AsyncHiredisParser",
@@ -17,4 +17,5 @@ __all__ = [
     "_HiredisParser",
     "_RESP2Parser",
     "_RESP3Parser",
+    "parse_url",
 ]
